perms for team perms

+++ index.js
@@ -35,8 +35,12 @@ client.on("message", async (msg) => {
 
   if (message.startsWith("!addtodo")) {
     const messageDetails = message.substr(message.indexOf(" ") + 1);
-    console.log(messageDetails);
-    addTodo(msg, messageDetails);
+    const team = messageDetails.split(" ")[0];
+    if (msg.member.roles.cache.some((r) => r.name.toLowerCase() === team)) {
+      addTodo(msg, messageDetails);
+    } else {
+      msg.reply("Access denied to command");
+    }
   }
 
   if (message.startsWith("!ideas")) {
@@ -49,7 +53,11 @@ client.on("message", async (msg) => {
 
   if (message.startsWith("!todo")) {
     const team = message.substr(message.indexOf(" ") + 1);
-    getTodo(msg, team, dsc);
+    if (msg.member.roles.cache.some((r) => r.name.toLowerCase() === team)) {
+      getTodo(msg, team, dsc);
+    } else {
+      msg.reply("Access denied to command");
+    }
   }
 
   if (message.startsWith("!removeidea")) {
@@ -63,7 +71,12 @@ client.on("message", async (msg) => {
 
   if (message.startsWith("!removetodo")) {
     const messageDetails = message.substr(message.indexOf(" ") + 1);
-    deleteTodo(msg, messageDetails);
+    const team = messageDetails.split(" ")[0];
+    if (msg.member.roles.cache.some((r) => r.name.toLowerCase() === team)) {
+      deleteTodo(msg, messageDetails);
+    } else {
+      msg.reply("Access denied to command");
+    }
   }
 
   if (message.startsWith("!rules")) {
